refactor: moved host and player components into /quiz
