+++ client/app.js
@@ -1,12 +1,8 @@
 import React from 'react'
-import firebase from 'firebase'
 import Routes from './routes'
 import {connect} from 'react-redux'
 import {withRouter} from 'react-router-dom'
 import {setPlayerThunk} from './store/user'
-
-// Front-end firebase is firebase.database()
-const database = firebase.database()
 
 class App extends React.Component {
   render() {
