+++ client/components/index.js
@@ -8,4 +8,4 @@ export {default as Welcome} from './Welcome'
 export {default as Join} from './Join'
 export {default as CreateGame} from './CreateGame'
 export {default as Room} from './Room'
-export {default as HostPlaying} from './HostPlaying'
+export {default as HostPlaying} from './quiz/HostPlaying'
